Consolidate preferences import and improve error handling
